+++ Source/PrefabricatorEditor/Private/Asset/PrefabricatorActorFactory.cpp
@@ -1,4 +1,4 @@
-//$ Copyright 2015-20, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
+//$ Copyright 2015-21, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
 
 #include "Asset/PrefabricatorActorFactory.h"
 
@@ -7,7 +7,7 @@
 #include "Prefab/PrefabComponent.h"
 #include "Prefab/PrefabTools.h"
 
-#include "AssetData.h"
+#include "AssetRegistry/AssetData.h"
 
 DEFINE_LOG_CATEGORY_STATIC(LogPrefabricatorActorFactory, Log, All);