+++ Source/PrefabricatorRuntime/PrefabricatorRuntime.Build.cs
@@ -21,6 +21,7 @@ namespace UnrealBuildTool.Rules
 					"CoreUObject",
 				    "Engine",
                     "PropertyPath",
+                    "DeveloperSettings",
 					// ... add other public dependencies that you statically link with here ...
 				}
 				);