+++ Source/PrefabricatorRuntime/Private/PrefabricatorSettings.cpp
@@ -1,12 +1,15 @@
-//$ Copyright 2015-20, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
+//$ Copyright 2015-21, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
 
 #include "PrefabricatorSettings.h"
 
+#include "Components/DrawFrustumComponent.h"
 
 UPrefabricatorSettings::UPrefabricatorSettings()
 {
 	//Do not change default behavior.
 	PivotPosition = EPrefabricatorPivotPosition::Center;
+
+	IgnoreBoundingBoxForObjects.Add(UDrawFrustumComponent::StaticClass());
 }
 
 /** Gets the settings container name for the settings, either Project or Editor */