+++ Source/PrefabricatorRuntime/Public/Prefab/Random/PrefabRandomizerActor.h
@@ -1,7 +1,9 @@
-//$ Copyright 2015-20, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
+//$ Copyright 2015-21, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
 
 #pragma once
 #include "CoreMinimal.h"
+#include "Prefab/PrefabActor.h"
+
 #include "GameFramework/Actor.h"
 #include "PrefabRandomizerActor.generated.h"
 
@@ -26,7 +28,7 @@ public:
 
 public:
 	UPROPERTY(EditAnywhere, Category = "Prefabricator")
-	bool bRandomizeOnBeginPlay = true;
+	bool bRandomizeOnBeginPlay = false;
 
 	UPROPERTY(EditAnywhere, Category = "Prefabricator")
 	int32 SeedOffset = 0;
@@ -39,6 +41,10 @@ public:
 
 	UPROPERTY(EditAnywhere, Category = "Prefabricator")
 	bool bFastSyncBuild = false;
+
+	/** If left empty, everything in the level will be randomized.  If set, only the actors in the list will be randomized. */
+	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Prefabricator")
+	TArray<APrefabActor*> ActorsToRandomize;
 
 private:
 	TSharedPtr<class FPrefabBuildSystem> BuildSystem;