+++ Source/PrefabricatorRuntime/Public/PrefabricatorSettings.h
@@ -1,4 +1,4 @@
-//$ Copyright 2015-20, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
+//$ Copyright 2015-21, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
 
 #pragma once
 #include "CoreMinimal.h"
@@ -24,11 +24,28 @@ class PREFABRICATORRUNTIME_API UPrefabricatorSettings : public UDeveloperSetting
 public:
 
 	/* Position pivot on extreme corner (bottom, left up), instead of center. */
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, config, Category = "Settings")
+	UPROPERTY(config, EditAnywhere, Category = "General Settings")
 	EPrefabricatorPivotPosition PivotPosition;
 
-	UPROPERTY(EditAnywhere, BlueprintReadOnly, config, Category = "Settings")
+	/** Whenever a prefab is saved, update all the other similar prefabs in the scene to reflect this new change */
+	UPROPERTY(config, EditAnywhere, Category = "General Settings")
 	bool bAllowDynamicUpdate = true;
+
+	/** Whenever a prefab is saved, update all the other similar prefabs in the scene to reflect this new change */
+	UPROPERTY(config, EditAnywhere, Category = "General Settings", Meta=(ConfigRestartRequired=true))
+	TSet<UClass*> IgnoreBoundingBoxForObjects;
+	
+	/** Use this angle while saving the prefab asset */
+	UPROPERTY(config, EditAnywhere, Category = "Thumbnail")
+	float DefaultThumbnailPitch = -11.25;
+
+	/** Use this angle while saving the prefab asset */
+	UPROPERTY(config, EditAnywhere, Category = "Thumbnail")
+	float DefaultThumbnailYaw = -157.5;
+
+	/** Use this zoom value while saving the prefab asset */
+	UPROPERTY(config, EditAnywhere, Category = "Thumbnail")
+	float DefaultThumbnailZoom = 0;
 
 public:
 	UPrefabricatorSettings();