+++ Source/PrefabricatorRuntime/Public/Utils/PrefabricatorFunctionLibrary.h
@@ -1,4 +1,4 @@
-//$ Copyright 2015-20, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
+//$ Copyright 2015-21, Code Respawn Technologies Pvt Ltd - All Rights Reserved $//
 
 #pragma once
 #include "CoreMinimal.h"
@@ -21,7 +21,15 @@ public:
 	static void RandomizePrefab(APrefabActor* PrefabActor, const FRandomStream& InRandom);
 
 	UFUNCTION(BlueprintCallable, Category = "Prefabricator")
+	static void UnlinkPrefab(APrefabActor* PrefabActor);
+
+	UFUNCTION(BlueprintCallable, Category = "Prefabricator")
 	static void GetAllAttachedActors(AActor* Prefab, TArray<AActor*>& AttachedActors);
 
+	UFUNCTION(BlueprintCallable, Category = "Prefabricator")
+	static void SetPrefabAsset(APrefabActor* PrefabActor, UPrefabricatorAssetInterface* Prefab, bool bReloadPrefab);
+
+	UFUNCTION(BlueprintCallable, Category = "Prefabricator")
+	static APrefabActor* FindTopMostPrefabActor(AActor* InActor);
 };